main/demos/retraining/registry/requirements.txt
flask_api
flask_caching
flask
pandas
joblib
sklearn
redis
google-cloud-storage
mlflow
